# backend/api/gen/siem/v1/feeds_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: siem/v1/feeds.proto

package siemv1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationFeedsCreateFeed = "/siem.v1.Feeds/CreateFeed"
const OperationFeedsDeleteFeed = "/siem.v1.Feeds/DeleteFeed"
const OperationFeedsGetFeed = "/siem.v1.Feeds/GetFeed"
const OperationFeedsGetFeedHealth = "/siem.v1.Feeds/GetFeedHealth"
const OperationFeedsListFeeds = "/siem.v1.Feeds/ListFeeds"
const OperationFeedsListRejectedEvents = "/siem.v1.Feeds/ListRejectedEvents"
const OperationFeedsTestFeed = "/siem.v1.Feeds/TestFeed"
const OperationFeedsUpdateFeed = "/siem.v1.Feeds/UpdateFeed"

type FeedsHTTPServer interface {
	// CreateFeed The credential is accepted here, stored in the secret manager, and never echoed.
	CreateFeed(context.Context, *CreateFeedRequest) (*Feed, error)
	// DeleteFeed Soft-disables the feed. Ingested data is never removed by this operation.
	DeleteFeed(context.Context, *DeleteFeedRequest) (*DeleteFeedResponse, error)
	GetFeed(context.Context, *GetFeedRequest) (*Feed, error)
	GetFeedHealth(context.Context, *GetFeedHealthRequest) (*FeedHealth, error)
	ListFeeds(context.Context, *ListFeedsRequest) (*ListFeedsResponse, error)
	// ListRejectedEvents Dead-lettered events, so a rejected delivery is visible rather than silent.
	ListRejectedEvents(context.Context, *ListRejectedEventsRequest) (*ListRejectedEventsResponse, error)
	TestFeed(context.Context, *TestFeedRequest) (*TestFeedResponse, error)
	UpdateFeed(context.Context, *UpdateFeedRequest) (*Feed, error)
}

func RegisterFeedsHTTPServer(s *http.Server, srv FeedsHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/feeds", _Feeds_ListFeeds0_HTTP_Handler(srv))
	r.POST("/api/v1/feeds", _Feeds_CreateFeed0_HTTP_Handler(srv))
	r.GET("/api/v1/feeds/{feed_id}", _Feeds_GetFeed0_HTTP_Handler(srv))
	r.PATCH("/api/v1/feeds/{feed_id}", _Feeds_UpdateFeed0_HTTP_Handler(srv))
	r.DELETE("/api/v1/feeds/{feed_id}", _Feeds_DeleteFeed0_HTTP_Handler(srv))
	r.POST("/api/v1/feeds/{feed_id}/test", _Feeds_TestFeed0_HTTP_Handler(srv))
	r.GET("/api/v1/feeds/{feed_id}/health", _Feeds_GetFeedHealth0_HTTP_Handler(srv))
	r.GET("/api/v1/feeds/{feed_id}/rejected", _Feeds_ListRejectedEvents0_HTTP_Handler(srv))
}

func _Feeds_ListFeeds0_HTTP_Handler(srv FeedsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListFeedsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFeedsListFeeds)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListFeeds(ctx, req.(*ListFeedsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListFeedsResponse)
		return ctx.Result(200, reply)
	}
}

func _Feeds_CreateFeed0_HTTP_Handler(srv FeedsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateFeedRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFeedsCreateFeed)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateFeed(ctx, req.(*CreateFeedRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Feed)
		return ctx.Result(200, reply)
	}
}

func _Feeds_GetFeed0_HTTP_Handler(srv FeedsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetFeedRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFeedsGetFeed)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetFeed(ctx, req.(*GetFeedRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Feed)
		return ctx.Result(200, reply)
	}
}

func _Feeds_UpdateFeed0_HTTP_Handler(srv FeedsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateFeedRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFeedsUpdateFeed)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateFeed(ctx, req.(*UpdateFeedRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Feed)
		return ctx.Result(200, reply)
	}
}

func _Feeds_DeleteFeed0_HTTP_Handler(srv FeedsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteFeedRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFeedsDeleteFeed)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteFeed(ctx, req.(*DeleteFeedRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteFeedResponse)
		return ctx.Result(200, reply)
	}
}

func _Feeds_TestFeed0_HTTP_Handler(srv FeedsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in TestFeedRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFeedsTestFeed)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.TestFeed(ctx, req.(*TestFeedRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*TestFeedResponse)
		return ctx.Result(200, reply)
	}
}

func _Feeds_GetFeedHealth0_HTTP_Handler(srv FeedsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetFeedHealthRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFeedsGetFeedHealth)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetFeedHealth(ctx, req.(*GetFeedHealthRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*FeedHealth)
		return ctx.Result(200, reply)
	}
}

func _Feeds_ListRejectedEvents0_HTTP_Handler(srv FeedsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListRejectedEventsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFeedsListRejectedEvents)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListRejectedEvents(ctx, req.(*ListRejectedEventsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListRejectedEventsResponse)
		return ctx.Result(200, reply)
	}
}

type FeedsHTTPClient interface {
	// CreateFeed The credential is accepted here, stored in the secret manager, and never echoed.
	CreateFeed(ctx context.Context, req *CreateFeedRequest, opts ...http.CallOption) (rsp *Feed, err error)
	// DeleteFeed Soft-disables the feed. Ingested data is never removed by this operation.
	DeleteFeed(ctx context.Context, req *DeleteFeedRequest, opts ...http.CallOption) (rsp *DeleteFeedResponse, err error)
	GetFeed(ctx context.Context, req *GetFeedRequest, opts ...http.CallOption) (rsp *Feed, err error)
	GetFeedHealth(ctx context.Context, req *GetFeedHealthRequest, opts ...http.CallOption) (rsp *FeedHealth, err error)
	ListFeeds(ctx context.Context, req *ListFeedsRequest, opts ...http.CallOption) (rsp *ListFeedsResponse, err error)
	// ListRejectedEvents Dead-lettered events, so a rejected delivery is visible rather than silent.
	ListRejectedEvents(ctx context.Context, req *ListRejectedEventsRequest, opts ...http.CallOption) (rsp *ListRejectedEventsResponse, err error)
	TestFeed(ctx context.Context, req *TestFeedRequest, opts ...http.CallOption) (rsp *TestFeedResponse, err error)
	UpdateFeed(ctx context.Context, req *UpdateFeedRequest, opts ...http.CallOption) (rsp *Feed, err error)
}

type FeedsHTTPClientImpl struct {
	cc *http.Client
}

func NewFeedsHTTPClient(client *http.Client) FeedsHTTPClient {
	return &FeedsHTTPClientImpl{client}
}

// CreateFeed The credential is accepted here, stored in the secret manager, and never echoed.
func (c *FeedsHTTPClientImpl) CreateFeed(ctx context.Context, in *CreateFeedRequest, opts ...http.CallOption) (*Feed, error) {
	var out Feed
	pattern := "/api/v1/feeds"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationFeedsCreateFeed))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteFeed Soft-disables the feed. Ingested data is never removed by this operation.
func (c *FeedsHTTPClientImpl) DeleteFeed(ctx context.Context, in *DeleteFeedRequest, opts ...http.CallOption) (*DeleteFeedResponse, error) {
	var out DeleteFeedResponse
	pattern := "/api/v1/feeds/{feed_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationFeedsDeleteFeed))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *FeedsHTTPClientImpl) GetFeed(ctx context.Context, in *GetFeedRequest, opts ...http.CallOption) (*Feed, error) {
	var out Feed
	pattern := "/api/v1/feeds/{feed_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationFeedsGetFeed))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *FeedsHTTPClientImpl) GetFeedHealth(ctx context.Context, in *GetFeedHealthRequest, opts ...http.CallOption) (*FeedHealth, error) {
	var out FeedHealth
	pattern := "/api/v1/feeds/{feed_id}/health"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationFeedsGetFeedHealth))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *FeedsHTTPClientImpl) ListFeeds(ctx context.Context, in *ListFeedsRequest, opts ...http.CallOption) (*ListFeedsResponse, error) {
	var out ListFeedsResponse
	pattern := "/api/v1/feeds"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationFeedsListFeeds))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListRejectedEvents Dead-lettered events, so a rejected delivery is visible rather than silent.
func (c *FeedsHTTPClientImpl) ListRejectedEvents(ctx context.Context, in *ListRejectedEventsRequest, opts ...http.CallOption) (*ListRejectedEventsResponse, error) {
	var out ListRejectedEventsResponse
	pattern := "/api/v1/feeds/{feed_id}/rejected"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationFeedsListRejectedEvents))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *FeedsHTTPClientImpl) TestFeed(ctx context.Context, in *TestFeedRequest, opts ...http.CallOption) (*TestFeedResponse, error) {
	var out TestFeedResponse
	pattern := "/api/v1/feeds/{feed_id}/test"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationFeedsTestFeed))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *FeedsHTTPClientImpl) UpdateFeed(ctx context.Context, in *UpdateFeedRequest, opts ...http.CallOption) (*Feed, error) {
	var out Feed
	pattern := "/api/v1/feeds/{feed_id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationFeedsUpdateFeed))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}