# backend/api/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/feeds:
        get:
            tags:
                - Feeds
            operationId: Feeds_ListFeeds
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListFeedsResponse'
        post:
            tags:
                - Feeds
            description: The credential is accepted here, stored in the secret manager, and never echoed.
            operationId: Feeds_CreateFeed
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateFeedRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Feed'
    /api/v1/feeds/{feedId}:
        get:
            tags:
                - Feeds
            operationId: Feeds_GetFeed
            parameters:
                - name: feedId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Feed'
        delete:
            tags:
                - Feeds
            description: Soft-disables the feed. Ingested data is never removed by this operation.
            operationId: Feeds_DeleteFeed
            parameters:
                - name: feedId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteFeedResponse'
        patch:
            tags:
                - Feeds
            operationId: Feeds_UpdateFeed
            parameters:
                - name: feedId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateFeedRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Feed'
...
    /api/v1/feeds/{feedId}/rejected:
        get:
            tags:
                - Feeds
            description: Dead-lettered events, so a rejected delivery is visible rather than silent.
            operationId: Feeds_ListRejectedEvents
            parameters:
                - name: feedId
                  in: path
                  required: true
                  schema:
                    type: string
                - name: range.from
                  in: query
                  schema:
                    type: string
                    format: date-time
                - name: range.to
                  in: query
                  schema:
                    type: string
                    format: date-time
                - name: reasonCode
                  in: query
                  schema:
                    type: string
                - name: page.cursor
                  in: query
                  description: Opaque cursor from a previous PageResponse. Empty requests the first page.
                  schema:
                    type: string
                - name: page.limit
                  in: query
                  description: Server caps this at QUERY_MAX_RESULT_ROWS regardless of what is requested.
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListRejectedEventsResponse'
    /api/v1/feeds/{feedId}/test:
        post:
            tags:
                - Feeds
            operationId: Feeds_TestFeed
            parameters:
                - name: feedId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TestFeedRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TestFeedResponse'
...
components:
    schemas:
...
        CreateFeedRequest:
            type: object
            properties:
                vendor:
                    enum:
                        - VENDOR_UNSPECIFIED
                        - VENDOR_CLOUDFLARE
                        - VENDOR_F5
                        - VENDOR_DATADOME
                    type: string
                    format: enum
                name:
                    type: string
                deliveryMode:
                    enum:
                        - DELIVERY_MODE_UNSPECIFIED
                        - DELIVERY_MODE_PUSH
                        - DELIVERY_MODE_PULL
                    type: string
                    format: enum
                credential:
                    type: string
                    description: Write-only. Supplied once, stored in the secret manager, never read back.
                signingSecret:
                    type: string
                    description: Optional HMAC key for vendors that sign their deliveries.
                pullConfig:
                    type: string
                quotaEventsPerSec:
                    type: integer
                    format: uint32
                quotaBytesPerDay:
                    type: string
...
        DeleteFeedResponse:
            type: object
            properties: {}
...
        Feed:
            type: object
            properties:
                feedId:
                    type: string
                vendor:
                    enum:
                        - VENDOR_UNSPECIFIED
                        - VENDOR_CLOUDFLARE
                        - VENDOR_F5
                        - VENDOR_DATADOME
                    type: string
                    format: enum
                name:
                    type: string
                deliveryMode:
                    enum:
                        - DELIVERY_MODE_UNSPECIFIED
                        - DELIVERY_MODE_PUSH
                        - DELIVERY_MODE_PULL
                    type: string
                    format: enum
                enabled:
                    type: boolean
                credentialConfigured:
                    type: boolean
                    description: True when a credential is configured. The credential itself is NEVER returned.
                signingConfigured:
                    type: boolean
                pullConfig:
                    type: string
                quotaEventsPerSec:
                    type: integer
                    format: uint32
                quotaBytesPerDay:
                    type: string
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
                health:
                    allOf:
                        - $ref: '#/components/schemas/FeedHealth'
                    description: Populated on list and get so the console can show status without a second call.
        FeedHealth:
            type: object
            properties:
                feedId:
                    type: string
                lastEventAt:
                    type: string
                    format: date-time
                silent:
                    type: boolean
                    description: True when nothing has arrived for longer than the expected interval.
                eventsPerSec:
                    type: number
                    format: double
                eventsReceived1h:
                    type: string
                eventsRejected1h:
                    type: string
                duplicatesSuppressed1h:
                    type: string
                bytesReceived1h:
                    type: string
                maxIngestLagMs:
                    type: integer
                    format: uint32
                credentialValid:
                    type: boolean
                schemaDriftWarning:
                    type: boolean
                    description: Set when unrecognized vendor fields crossed the drift threshold (FR-012).
                driftFields:
                    type: array
                    items:
                        type: string
            description: |-
                FeedHealth answers "is this feed working" (FR-008).

                 Silence detection lives here because a feed that stopped sending looks exactly like
                 clean traffic on a dashboard. `silent` is what distinguishes them.
...
        ListFeedsResponse:
            type: object
            properties:
                feeds:
                    type: array
                    items:
                        $ref: '#/components/schemas/Feed'
        ListRejectedEventsResponse:
            type: object
            properties:
                events:
                    type: array
                    items:
                        $ref: '#/components/schemas/RejectedEvent'
                page:
                    $ref: '#/components/schemas/PageResponse'
...
        PageResponse:
            type: object
            properties:
                nextCursor:
                    type: string
                    description: Empty when there are no further pages.
                total:
                    type: string
                    description: |-
                        Approximate match count. Exact counts over hundreds of millions of rows cannot
                         meet the latency budget, so the API states when a count is an estimate rather
                         than implying precision it does not have.
                totalIsEstimate:
                    type: boolean
...
        RejectedEvent:
            type: object
            properties:
                rejectedAt:
                    type: string
                    format: date-time
                vendor:
                    enum:
                        - VENDOR_UNSPECIFIED
                        - VENDOR_CLOUDFLARE
                        - VENDOR_F5
                        - VENDOR_DATADOME
                    type: string
                    format: enum
                reasonCode:
                    type: string
                reasonDetail:
                    type: string
                payload:
                    type: string
                    description: |-
                        The offending payload, so an operator can see exactly what the vendor sent.
                         Rendered as inert text by the console.
                batchId:
                    type: string
...
        TestFeedRequest:
            type: object
            properties:
                feedId:
                    type: string
        TestFeedResponse:
            type: object
            properties:
                reachable:
                    type: boolean
                credentialValid:
                    type: boolean
                detail:
                    type: string
                    description: Human-readable outcome. Never contains the credential or an internal address.
...
        UpdateFeedRequest:
            type: object
            properties:
                feedId:
                    type: string
                name:
                    type: string
                enabled:
                    type: boolean
                credential:
                    type: string
                    description: Rotates the credential. Omitted leaves the existing one in place.
                signingSecret:
                    type: string
                pullConfig:
                    type: string
                quotaEventsPerSec:
                    type: integer
                    format: uint32
                quotaBytesPerDay:
                    type: string
...